requirements.txt
flask==3.0.0
line-bot-sdk==3.5.0
google-generativeai==0.3.1
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
